REmoving unused classes
